eChatBot - PLAYER_MESSAGE_TRIGGERS_SILENCED_CLEAR_ON_SILENCE Updated timer used for PLAYER_WATCH_SILENCED ERROR_SHOW_MESSAGE_BOX - Determines if message box containing error is shown prior to closing due to an error on windows PLAYER_WATCH_SERVER_DISCONNECTED - Menu level check
Refactor / fixes

+++ src/engine/ePlayer.h
@@ -369,9 +369,9 @@ public:
     static int nameSpeakPlayerID;
     static int playerUpdateIteration;
 
-    static REAL nextSpeakTimePrefix;
+    static double nextSpeakTimePrefix;
     static tString nextSpeakTimePrefixCommonPrefix;
-    static REAL nextSpeakTime;
+    static double nextSpeakTime;
 
     tString lastSentName;
